Create README.md
This is a trivial "working" applications designed with Spring Boot that manages reservations. I have coded it mainly to create a reference for myself (and anybody interested in Spring Boot). 

What is Spring Boot?
As stated by its inventors, Spring Boot makes it easy to create Spring-powered, production-grade applications and services with absolute minimum fuss.